roles/daemon/split bootstrap (#39)
* feat: Add download and extraction steps for daemon bootstrap file

- Download the daemon bootstrap file from the specified URL and store it in /tmp directory
- Extract the downloaded file to the specified daemon path
- Update daemon variable to include the name of the bootstrap file
- Minor formatting changes in the code

This commit adds the necessary steps to download and extract the daemon bootstrap file, which is required for th
e daemon to function properly. The bootstrap file is downloaded from the specified URL and stored in the /tmp di
rectory. Then, the file is extracted to the specified daemon path. Additionally, the daemon variable is updated
to include the name of the bootstrap file.

* Update hosts.ini and daemon-bootstrap.yml, add bootstrap_file to daemon vars

Details:
- Update hosts.ini to add a new group 'dufus' with two hosts
- Update daemon-bootstrap.yml to download the bootstrap file to /tmp and extract it from there
- Add 'bootstrap_file' variable to daemon vars in vars.yml